chore(landing): move create board button to its own component
