+++ next-tavla/src/Admin/scenarios/Landing/index.tsx
@@ -1,9 +1,5 @@
-import { Button } from '@entur/button'
 import { Heading1, ListItem, Paragraph, UnorderedList } from '@entur/typography'
 import classes from './styles.module.css'
-import { addBoardSettings } from 'utils/firebase'
-import { useRouter } from 'next/router'
-import { useState } from 'react'
 import Image from 'next/image'
 import landingImage from 'assets/illustrations/Tavla-illustration.png'
 import tavla from 'assets/illustrations/Tavla-screenshot.png'
@@ -11,17 +7,6 @@ import { Contrast } from '@entur/layout'
 import classNames from 'classnames'
 
 function Landing() {
-    const router = useRouter()
-    const [loading, setLoading] = useState(false)
-
-    async function handleCreateNewBoard() {
-        setLoading(true)
-        const createdBoard = await addBoardSettings({
-            tiles: [],
-        })
-        await router.push('/edit/' + createdBoard.id)
-    }
-
     return (
         <div className={classes.container}>
             <Contrast className={classes.centeredContainer}>
@@ -30,17 +15,6 @@ function Landing() {
                     <Heading1 className={classes.subheading}>
                         for reisende
                     </Heading1>
-
-                    <Button
-                        onClick={handleCreateNewBoard}
-                        variant="primary"
-                        disabled={loading}
-                        loading={loading}
-                        width="fluid"
-                        className={classes.button}
-                    >
-                        Opprett ny tavle
-                    </Button>
                 </div>
                 <div className={classNames(classes.content, classes.topImage)}>
                     <Image
